Moved store to RecordingService
