+++ lib/Service/RecordingService.php
@@ -1,0 +1,128 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * @copyright Copyright (c) 2022 Vitor Mattos <vitor@php.rio>
+ *
+ * @author Vitor Mattos <vitor@php.rio>
+ *
+ * @license GNU AGPL version 3 or any later version
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+namespace OCA\Talk\Service;
+
+use InvalidArgumentException;
+use OC\Files\Filesystem;
+use OC\User\NoUserException;
+use OCA\Talk\Config;
+use OCA\Talk\Exceptions\ParticipantNotFoundException;
+use OCA\Talk\Room;
+use OCP\Files\Folder;
+use OCP\Files\IMimeTypeDetector;
+use OCP\Files\IRootFolder;
+use OCP\Files\NotFoundException;
+use OCP\Files\NotPermittedException;
+
+class RecordingService {
+	private IMimeTypeDetector $mimeTypeDetector;
+	private ParticipantService $participantService;
+	private Config $config;
+
+	public function __construct(
+		IMimeTypeDetector $mimeTypeDetector,
+		ParticipantService $participantService,
+		Config $config
+	) {
+		$this->mimeTypeDetector = $mimeTypeDetector;
+		$this->participantService = $participantService;
+		$this->config = $config;
+	}
+
+	public function store(Room $room, string $owner, array $file, IRootFolder $rootFolder): void {
+		if (
+			$file['error'] !== 0 ||
+			!is_uploaded_file($file['tmp_name']) ||
+			Filesystem::isFileBlacklisted($file['tmp_name'])
+		) {
+			throw new InvalidArgumentException('invalid_file');
+		}
+
+		$content = file_get_contents($file['tmp_name']);
+		unlink($file['tmp_name']);
+
+		$mimeType = $this->mimeTypeDetector->detectString($content);
+		$allowedMimeTypes = [
+			'video/mp4',
+			'video/mpeg',
+			'video/ogg',
+			'audio/mp3',
+			'audio/ogg',
+		];
+		if (!in_array($mimeType, $allowedMimeTypes)) {
+			throw new InvalidArgumentException('file_mimetype');
+		}
+
+		$recordFileName = escapeshellcmd($file['name']);
+		if (strlen($recordFileName) !== strlen($file['name'])) {
+			throw new InvalidArgumentException('file_name');
+		}
+
+		$extensionFromMime = pathinfo(str_replace('/', '.', $mimeType), PATHINFO_EXTENSION);
+		$extensionFromFileName = strtolower(pathinfo($recordFileName, PATHINFO_EXTENSION));
+		if ($extensionFromFileName !== $extensionFromMime) {
+			throw new InvalidArgumentException('file_extension');
+		}
+
+		try {
+			$this->participantService->getParticipant($room, $owner);
+		} catch (ParticipantNotFoundException $e) {
+			throw new InvalidArgumentException('owner_participant');
+		}
+
+		try {
+			$recordingFolder = $this->getRecordingFolder($rootFolder, $owner, $room->getToken());
+			$recordingFolder->newFile($recordFileName, $content);
+		} catch (NoUserException $e) {
+			throw new InvalidArgumentException('owner_invalid');
+		} catch (NotPermittedException $e) {
+			throw new InvalidArgumentException('owner_permission');
+		}
+	}
+
+	private function getRecordingFolder(IRootFolder $rootFolder, string $owner, string $token): Folder {
+		$attachmentFolderName = $this->config->getAttachmentFolder($owner);
+
+		$userFolder = $rootFolder->getUserFolder($owner);
+		try {
+			$attachmentFolder = $userFolder->get($attachmentFolderName);
+		} catch (NotFoundException $e) {
+			$attachmentFolder = $userFolder->newFolder($attachmentFolderName);
+		}
+		$recordingRootFolderName = $this->config->getRecordingFolder($owner);
+		try {
+			$recordingRootFolder = $attachmentFolder->get($recordingRootFolderName);
+		} catch (NotFoundException $e) {
+			$recordingRootFolder = $attachmentFolder->newFolder($recordingRootFolderName);
+		}
+		try {
+			$recordingFolder = $recordingRootFolder->get($token);
+		} catch (NotFoundException $e) {
+			$recordingFolder = $recordingRootFolder->newFolder($token);
+		}
+		return $recordingFolder;
+	}
+}
